EAMxx: renamed field_impl to better reflect its content
