+++ components/eamxx/src/share/field/tests/field_tests.cpp
@@ -1,13 +1,13 @@
 #include <catch2/catch.hpp>
-#include <numeric>
 
 #include "share/field/field_identifier.hpp"
 #include "share/field/field.hpp"
 #include "share/field/field_utils.hpp"
-#include "share/field/field_impl.hpp"
 #include "share/core/eamxx_setup_random_test.hpp"
 
 #include <ekat_pack.hpp>
+
+#include <numeric>
 
 namespace {
