+++ components/eamxx/src/share/field/tests/field_update_tests.cpp
@@ -4,7 +4,6 @@
 #include "share/field/field_identifier.hpp"
 #include "share/field/field.hpp"
 #include "share/field/field_utils.hpp"
-#include "share/field/field_impl.hpp"
 #include "share/core/eamxx_setup_random_test.hpp"
 
 namespace {
